662: extract common match pattern code
