+++ src/devTools/editor/extensionPoints/base.ts
@@ -23,7 +23,6 @@ import {
   ReaderFormState,
   ReaderReferenceFormState,
 } from "@/devTools/editor/editorSlice";
-import psl, { ParsedDomain } from "psl";
 import { castArray, identity, isPlainObject } from "lodash";
 import brickRegistry from "@/blocks/registry";
 import { ReaderConfig, ReaderReference } from "@/blocks/readers/factory";
@@ -37,6 +36,7 @@ import {
 } from "@/extensionPoints/types";
 import { find as findBrick } from "@/registry/localRegistry";
 import React from "react";
+import { defaultMatchPattern, getDomain } from "@/permissions/patterns";
 
 export interface WizardStep {
   step: string;
@@ -46,26 +46,6 @@ export interface WizardStep {
     available?: boolean;
   }>;
   extraProps?: Record<string, unknown>;
-}
-
-function getPathFromUrl(url: string): string {
-  return url.split("?")[0];
-}
-
-function defaultMatchPattern(url: string): string {
-  const cleanURL = getPathFromUrl(url);
-  console.debug(`Clean URL: ${cleanURL}`);
-  const obj = new URL(cleanURL);
-  // https://developer.mozilla.org/en-US/docs/Web/API/URLSearchParams/entries
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any -- TypeScript definitions are incorrect
-  for (const [name] of (obj.searchParams as any).entries()) {
-    console.debug(`Deleting param ${name}`);
-    obj.searchParams.delete(name);
-  }
-  obj.pathname = "*";
-  obj.hash = "";
-  console.debug(`Generate match pattern`, { href: obj.href });
-  return obj.href;
 }
 
 function defaultReader(frameworks: FrameworkMeta[]): Framework {
@@ -143,12 +123,6 @@ export function makeBaseState(
     extension: {},
     extensionPoint: {},
   };
-}
-
-export function getDomain(url: string): string {
-  const urlClass = new URL(url);
-  const { domain } = psl.parse(urlClass.host.split(":")[0]) as ParsedDomain;
-  return domain;
 }
 
 export async function generateExtensionPointMetadata(
